+++ packages/argon-router-core/lib/index.ts
@@ -1,11 +1,13 @@
 export { createRoute } from './create-route';
 export { createRouter } from './create-router';
 export { chainRoute } from './chain-route';
+export { createRouterControls } from './create-router-controls';
 export { parameters } from './const';
 export type {
   Route,
   Router,
   Query,
+  OpenPayloadBase,
   RouteOpenedPayload,
   NavigatePayload,
   QueryTracker,