+++ packages/argon-router-paths/tests/index.test.ts
@@ -35,8 +35,33 @@ describe('parse path', () => {
     });
   });
 
+  test('parse path with generic parameter with dummy spaces (number)', () => {
+    const { parse } = compile('/profile/:id< number >');
+
+    expect(parse('/profile/12323')).toStrictEqual({
+      path: '/profile/12323',
+      params: { id: 12323 },
+    });
+  });
+
   test('parse path with generic parameter (union)', () => {
     const { parse } = compile('/profile/:id<hello|world>');
+
+    expect(parse('/profile/hello')).toStrictEqual({
+      path: '/profile/hello',
+      params: { id: 'hello' },
+    });
+
+    expect(parse('/profile/world')).toStrictEqual({
+      path: '/profile/world',
+      params: { id: 'world' },
+    });
+
+    expect(parse('/profile/test')).toStrictEqual(null);
+  });
+
+  test('parse path with generic parameter with dummy spaces (union)', () => {
+    const { parse } = compile('/profile/:id<hello | world >');
 
     expect(parse('/profile/hello')).toStrictEqual({
       path: '/profile/hello',
@@ -286,6 +311,26 @@ describe('parse path', () => {
     expect(parse('/profile')).toStrictEqual({
       path: '/profile',
       params: { id: undefined },
+    });
+  });
+
+  test('parse path without parameter but with extra parts', () => {
+    const fp = compile('/profile');
+    const sp = compile('/profile/:id');
+
+    expect(fp.parse('/profile/1')).toStrictEqual(null);
+    expect(sp.parse('/profile/1')).toStrictEqual({
+      path: '/profile/1',
+      params: { id: '1' },
+    });
+  });
+
+  test('parse path with array parameter and extra parts', () => {
+    const fp = compile('/items/:id{2,2}/:hi');
+
+    expect(fp.parse('/items/1/2/hello')).toStrictEqual({
+      path: '/items/1/2/hello',
+      params: { id: ['1', '2'], hi: 'hello' },
     });
   });
 });