+++ packages/argon-router-react/lib/index.ts
@@ -3,7 +3,7 @@ export { Link } from './link';
 export { createRouteView } from './create-route-view';
 export { createRoutesView } from './create-routes-view';
 export { createLazyRouteView } from './create-lazy-route-view';
-export { useRouter } from './use-router';
+export { useRouter, useRouterContext } from './use-router';
 export type {
   LinkProps,
   CreateLazyRouteViewProps,