+++ packages/argon-router-react/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.3](https://github.com/movpushmov/argon-router/compare/v0.8.2...v0.8.3) (2025-05-27)
+
+**Note:** Version bump only for package @argon-router/react
+
+## [0.8.2](https://github.com/movpushmov/argon-router/compare/v0.8.0...v0.8.2) (2025-05-27)
+
+### Bug Fixes
+
+- chained routes ([5d77fe5](https://github.com/movpushmov/argon-router/commit/5d77fe5ff637a816975482c2902458707a68195e))
+- create routes view ([a267e31](https://github.com/movpushmov/argon-router/commit/a267e313629fc62174407fe72bbc3e77f737f7b8))
+- update route view types ([ac02e7e](https://github.com/movpushmov/argon-router/commit/ac02e7e4078248f542c239366f9f376d213d4004))
+
+## [0.8.1](https://github.com/movpushmov/argon-router/compare/v0.8.0...v0.8.1) (2025-05-22)
+
+### Bug Fixes
+
+- update route view types ([ac02e7e](https://github.com/movpushmov/argon-router/commit/ac02e7e4078248f542c239366f9f376d213d4004))
+
+# [0.8.0](https://github.com/movpushmov/argon-router/compare/v0.7.0...v0.8.0) (2025-05-21)
+
+### Features
+
+- export useRouterContext ([599ab39](https://github.com/movpushmov/argon-router/commit/599ab39e258b9b05726e930d44c3589b2b7805f9))
+
+# [0.7.0](https://github.com/movpushmov/argon-router/compare/v0.6.3...v0.7.0) (2025-05-06)
+
+**Note:** Version bump only for package @argon-router/react
+
+## [0.6.3](https://github.com/movpushmov/argon-router/compare/v0.6.2...v0.6.3) (2025-04-30)
+
+**Note:** Version bump only for package @argon-router/react
+
+## [0.6.2](https://github.com/movpushmov/argon-router/compare/v0.6.1...v0.6.2) (2025-04-09)
+
+### Bug Fixes
+
+- react types ([82b2eac](https://github.com/movpushmov/argon-router/commit/82b2eace4e8ce5248e26babf3e9ac432cb3ce286))
+
 ## [0.6.1](https://github.com/movpushmov/argon-router/compare/v0.6.0...v0.6.1) (2025-03-26)
 
 **Note:** Version bump only for package @argon-router/react